refactor: adopt madara-style crates workspace, clean up workspace dependencies
