--- core/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "kakarot_rpc_core"
-version = "0.1.0"
-edition.workspace = true
-description.workspace = true
-homepage.workspace = true
-
-[dependencies]
-async-trait = "0.1.58"
-eyre = "0.6.8"
-log = "0.4.17"
-env_logger = "0.10.0"
-starknet = { git = "https://github.com/xJonathanLEI/starknet-rs" }
-url = "2.3.1"
-thiserror = "1.0.38"
-anyhow = "1.0.68"
-jsonrpsee = "0.16.2"
-reqwest = "0.11.13"
-hex = "0.4"
-reth-rpc-api = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-rpc-types = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-primitives = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-rlp = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-
-
-tokio = { version = "1.21.2", features = ["macros"] }
-serde_json = "1.0"
-serde = { version = "1.0" }
-futures = "0.3.26"
-
-wiremock = "0.5.17"
