--- core/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod client;
-pub mod helpers;
-pub mod utils;
