+++ crates/core/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "kakarot_rpc_core"
+version = "0.1.0"
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+
+[dependencies]
+async-trait = "0.1.58"
+eyre = "0.6.8"
+log = "0.4.17"
+env_logger = "0.10.0"
+url = "2.3.1"
+thiserror = "1.0.38"
+anyhow = "1.0.68"
+jsonrpsee = "0.16.2"
+reqwest = "0.11.13"
+hex = "0.4"
+reth-rpc-api.workspace = true
+reth-rpc-types.workspace = true
+reth-primitives.workspace = true
+reth-rlp.workspace = true
+starknet.workspace = true
+
+tokio = { version = "1.21.2", features = ["macros"] }
+serde_json = "1.0"
+serde = { version = "1.0" }
+futures = "0.3.26"
+
+wiremock = "0.5.17"
