+++ crates/core/src/client/client_api.rs
@@ -17,9 +17,8 @@ use starknet::{
 
 use thiserror::Error;
 extern crate hex;
-use crate::helpers::MaybePendingStarknetBlock;
 
-use crate::client::types::TokenBalances;
+use crate::client::{helpers::MaybePendingStarknetBlock, types::TokenBalances};
 use async_trait::async_trait;
 use reth_rpc_types::Index;
