+++ crates/core/src/client/helpers.rs
