+++ crates/core/src/mock/data/blocks/pending/starknet_getBlockWithTxHashes_pending.json
