+++ crates/core/src/mock/data/blocks/pending/starknet_getBlockWithTxs_pending.json
