+++ crates/core/src/mock/data/blocks/starknet_getBlockWithTxHashes.json
