+++ crates/core/src/mock/data/starknet_getCode.json
