+++ crates/core/src/mock/data/transactions/pending/starknet_getTransactionByBlockIdAndIndex.json
