+++ crates/core/src/mock/data/transactions/starknet_getClassHashAt.json
