+++ crates/core/src/mock/data/transactions/starknet_getTransactionByBlockIdAndIndex.json
