+++ crates/core/src/mock/data/transactions/starknet_getTransactionByHash_Declare.json
