+++ crates/core/src/mock/data/transactions/starknet_getTransactionByHash_Invoke.json
