+++ crates/core/src/mock/data/transactions/starknet_getTransactionByHash_L1Handler.json
