+++ crates/core/src/mock/data/transactions/starknet_getTransactionReceipt_Declare.json
