+++ crates/core/src/mock/data/transactions/starknet_getTransactionReceipt_DeployAccount.json
