+++ crates/core/src/mock/data/transactions/starknet_getTransactionReceipt_Invoke.json
