+++ crates/core/src/mock/data/transactions/starknet_getTransactionReceipt_L1Handler.json
