+++ crates/core/src/mock/wiremock_utils.rs
@@ -1,4 +1,4 @@
-use crate::helpers::ethers_block_id_to_starknet_block_id;
+use crate::client::helpers::ethers_block_id_to_starknet_block_id;
 use reqwest::StatusCode;
 use reth_primitives::{BlockId, H256};
 use serde::{Deserialize, Serialize};
