+++ crates/core/tests/client.rs
