+++ crates/rpc/Cargo.toml
@@ -24,13 +24,13 @@ serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0.38"
 hex = "0.4"
-reth-rpc-api = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-rpc-types = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-rlp = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
-reth-primitives = { git = "https://github.com/paradigmxyz/reth.git", rev = "8094ab18dbb230f177babceeb3d72c7a4c424490" }
+reth-rpc-api.workspace = true
+reth-rpc-types.workspace = true
+reth-rlp.workspace = true
+reth-primitives.workspace = true
 tracing = "0.1.34"
 tracing-subscriber = { version = "0.3.3", features = ["env-filter"] }
-starknet = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet.workspace = true
 reqwest = "0.11.13"
 dotenv = "0.15.0"
 mockall = "0.11.3"
