+++ crates/rpc/tests/serde.rs
