+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2022-12-15
+nightly-2023-06-18
